# reset timer in auto allocate benchmark