@ramaraochavali 2023-06-22
With Reset
=========
pkg: istio.io/istio/pilot/pkg/serviceregistry/serviceentry
BenchmarkAutoAllocateIPs
BenchmarkAutoAllocateIPs-10    	       4	 296139781 ns/op	22626758 B/op	  304024 allocs/op
PASS
ok  	istio.io/istio/pilot/pkg/serviceregistry/serviceentry	3.703s

Without Reset
-==========
goos: darwin
goarch: arm64
pkg: istio.io/istio/pilot/pkg/serviceregistry/serviceentry
BenchmarkAutoAllocateIPs
BenchmarkAutoAllocateIPs-10    	       3	 337533514 ns/op	32413349 B/op	  385160 allocs/op
PASS
ok  	istio.io/istio/pilot/pkg/serviceregistry/serviceentry	3.170s